(ts) @hickory/hash
