+++ packages/hash/.gitignore
@@ -1,0 +1,4 @@
+.rpt2_cache
+
+dist
+types
