+++ packages/hash/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.0-beta.3
+
+* Switch to TypeScript.
+
 ## 1.0.0-beta.2
 
 * Add `raw` option to `Hash` options. This defaults to a function that will generate the encoded version of a pathname (by setting an anchor element's `href` attribute and reading its `pathname`).
