+++ packages/hash/package.json
@@ -1,16 +1,18 @@
 {
   "name": "@hickory/hash",
-  "version": "1.0.0-beta.2",
+  "version": "1.0.0-beta.3",
   "description": "Hickory's hash history",
   "main": "dist/hickory-hash.common.js",
   "module": "dist/hickory-hash.es.js",
+  "types": "types/index.d.ts",
   "files": [
     "dist",
+    "types",
     "LICENSE",
     "README.md"
   ],
   "scripts": {
-    "prebuild": "rimraf dist",
+    "prebuild": "rimraf dist && rimraf types",
     "build": "node ./scripts/build",
     "prepublishOnly": "npm run build",
     "test": "node ./scripts/test"
@@ -27,18 +29,14 @@
   "author": "Paul Sherman",
   "license": "MIT",
   "dependencies": {
-    "@hickory/dom-utils": "^1.0.0-beta.0",
-    "@hickory/location-utils": "^1.0.0-beta.0",
-    "@hickory/root": "^1.0.0-beta.2"
+    "@hickory/dom-utils": "^1.0.0-beta.1",
+    "@hickory/location-utils": "^1.0.0-beta.1",
+    "@hickory/root": "^1.0.0-beta.3"
   },
   "devDependencies": {
-    "babel-core": "^6.25.0",
-    "babel-jest": "^20.0.3",
-    "babel-plugin-external-helpers": "^6.22.0",
-    "babel-plugin-transform-object-assign": "^6.22.0",
-    "babel-preset-env": "^1.6.0",
+    "@types/node": "^8.0.28",
     "jasmine": "^2.7.0",
-    "jest": "^20.0.4",
+    "jest": "^21.1.0",
     "karma": "^1.7.0",
     "karma-browserstack-launcher": "^1.3.0",
     "karma-chrome-launcher": "^2.2.0",
@@ -47,15 +45,29 @@
     "qs": "^6.5.0",
     "rimraf": "^2.6.1",
     "rollup": "^0.45.2",
-    "rollup-plugin-babel": "^2.7.1",
     "rollup-plugin-node-resolve": "^3.0.0",
     "rollup-plugin-replace": "^1.1.1",
-    "rollup-plugin-uglify": "^2.0.1"
+    "rollup-plugin-typescript2": "^0.5.2",
+    "rollup-plugin-uglify": "^2.0.1",
+    "ts-jest": "^21.0.1",
+    "typescript": "^2.5.2"
   },
   "jest": {
     "testEnvironment": "node",
+    "moduleFileExtensions": [
+      "ts",
+      "js"
+    ],
     "testMatch": [
-      "**/tests/unit/**/*.spec.js"
-    ]
+      "**/tests/unit/**/*.spec.ts"
+    ],
+    "transform": {
+      "\\.ts$": "<rootDir>/node_modules/ts-jest/preprocessor.js"
+    },
+    "globals": {
+      "ts-jest": {
+        "module": "es6"
+      }
+    }
   }
 }
