+++ packages/hash/src/hashTypes.ts
@@ -8,14 +8,19 @@ const DEFAULT = 'default';
 const CLEAN = 'clean';
 const BANG = 'bang';
 
+export interface EncodingFns {
+  encode: (path: string) => string;
+  decode: (path: string) => string;
+}
+
 // no matter with type of hash configuration we are using,
 // our decode function should return a string that begins
 // with a forward slash
-export default function hashEncoderAndDecoder(type = DEFAULT) {
+export default function hashEncoderAndDecoder(type: string = DEFAULT): EncodingFns {
   switch (type) {
   case CLEAN:
     return {  
-      encode: (path) => {
+      encode: (path: string): string => {
         // When we are at the root (/), we need to include the leading
         // slash because the hash needs at least one character after the
         // pound sign. We do the same even if there is a query or hash 
@@ -26,18 +31,18 @@ export default function hashEncoderAndDecoder(type = DEFAULT) {
         }
         return '#' + noSlash;
       },
-      decode: (path) => {
+      decode: (path: string): string => {
         const noHash = stripPrefix(path, '#');
         return ensureBeginsWith(noHash, '/');
       }
     };
   case BANG:
     return {
-      encode: (path) => {
+      encode: (path: string): string => {
         const withSlash = ensureBeginsWith(path, '/');
         return '#' + ensureBeginsWith(withSlash, '!');
       },
-      decode: (path) => {
+      decode: (path: string): string => {
         const noHash = stripPrefix(path, '#');
         const noBang = stripPrefix(noHash, '!');
         return ensureBeginsWith(noBang, '/');
@@ -46,10 +51,10 @@ export default function hashEncoderAndDecoder(type = DEFAULT) {
   case DEFAULT:
   default:
     return {
-      encode: (path) => {
+      encode: (path: string): string => {
         return '#' + ensureBeginsWith(path, '/');
       },
-      decode: (path) => {
+      decode: (path: string): string => {
         const noHash = stripPrefix(path, '#');
         return ensureBeginsWith(noHash, '/');
       }
